requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography, signxml
certifi==2018.10.15       # via requests, signxml
cffi==1.14.4              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via flask
cryptography==2.3.1       # via pyopenssl, signxml
defusedxml==0.5.0         # via signxml
eight==0.4.2              # via signxml
faker==2.0.3              # via -r requirements.in
flask-admin==1.5.1        # via -r requirements.in
flask-sqlalchemy==2.3.2   # via -r requirements.in
flask==1.0.2              # via -r requirements.in, flask-admin, flask-sqlalchemy
future==0.16.0            # via eight
idna==2.7                 # via cryptography, requests
importlib-resources==1.0.1  # via -r requirements.in
itsdangerous==1.1.0       # via flask
jinja2==2.10.3            # via flask
lxml==4.2.3               # via -r requirements.in, signxml
markupsafe==1.1.1         # via jinja2
psycopg2==2.8.6           # via -r requirements.in
pycparser==2.19           # via cffi
pyopenssl==17.5.0         # via signxml
python-codicefiscale==0.3.5
python-dateutil==2.7.5    # via faker, python-codicefiscale
python-slugify==3.0.6
pyyaml==5.1
requests==2.20.0
signxml==2.5.2
six==1.11.0               # via cryptography, faker, pyopenssl, python-dateutil
sqlalchemy==1.2.12
text-unidecode==1.3       # via faker, python-slugify
urllib3==1.24.3           # via requests
voluptuous==0.11.5
werkzeug==0.16.1          # via flask
wheel==0.33.6             # via importlib-resources
wtforms==2.2.1            # via flask-admin
signxml==2.5.2
flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
importlib-resources==1.0.1
lxml==4.2.3
pyyaml==5.1
python-codicefiscale==0.3.5
ujson==4.0.1
python3-saml==1.9.0
# The following packages are considered to be unsafe in a requirements file:
# setuptools
